web-scraping/requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
pinecone-client==2.0.0
openai==0.11.4
schedule==1.1.0
